update to use Paul's options, plus clean up
