+++ requirements.txt
@@ -1,0 +1,1 @@
+https://api.github.com/repos/GlobalFishingWatch/pipe-tools/tarball/14-logging-options
